Fixed blockhammer includes
