--- src/dram_controller/impl/plugin/bliss/bliss.h
@@ -1,15 +1,0 @@
-#ifndef RAMULATOR_PLUGIN_BLISS_H_
-#define RAMULATOR_PLUGIN_BLISS_H_
-
-namespace Ramulator
-{
-
-class IBLISS
-{
-  public:
-    virtual bool is_blacklisted(int source_id) = 0;
-};
-
-} // namespace Ramulator
-
-#endif // RAMULATOR_PLUGIN_BLISS_H_
